==> code/Stage1_clean_calendar_df.Rmd <==
---
title: "calendar_df"
author: "KarinaJonina"
date: "5/1/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


## Pupose of this code:
There are 9 million rows of data, which made it very difficult to rerun and test the code. My personal computer has 8GB of RAM and found it difficult to handle this amount of rows. Therefore, it was decided to create new dataset to summarise what is necessary from the dataset, so that the number of rows was reduced and the computer would not have so much difficulty with the rows.  

Two Rmd files were submitted, one analysing AirBnB data, the other treating and extracting important information from 'calendar.csv'. 

THIS RMD FILE TREATS CALENDAR.CSV

## Files submitted
Due to the 9 million rows in calendar.csv, it was decided to treat this dataset outside of this analysis due to the size of it. 

- calendar_df.Rmd - Code to treat calendar.csv and create the two datasets below to analyse in this report
- summarise_df.csv - A new dataset from calendar.csv  was created, where listings were examined by availability for each month within a year (26th June 2024 to 28th June 2025).
- available_nights_df.csv - A new dataset summarising the number of available nights for each listing in a year (26th June 2024 to 28th June 2025).
- calendar_df.html - Document presents calendar_df.Rmd in HTML


```{r echo=FALSE, warning=FALSE, message=FALSE}
library(tidyverse)
library(ggplot2)
library(dplyr)
library(stringr)
library(knitr)
library(plotly)
library(forcats)
library(crosstalk)
library(here)



options(scipen = 999)
```

## Dataset 'calendar_df'
To examine next year's availability, the 'calendar.csv' for Ireland was downloaded from the AirBnB website. It was downloaded from http://insideairbnb.com/get-the-data.html (calendar.csv).  This dataset is between 26th June 2024 to 28th June 2025. This dataset states whether the AirBnB listing is available on a specific date.

Data Dictionary was summarised from https://docs.google.com/spreadsheets/d/1iWCNJcSutYqpULSQHlNyGInUvHg2BoUGoNRIGa6Szc4/edit#gid=982310896

```{r echo=FALSE, warning=FALSE}

# Make sure to change this so that you can rerun code with 2021 data
DIRECTORY = "data_2024"

# specifying the path to the CSV file
file_path <- here(DIRECTORY, "calendar.csv")

# Read the CSV file
calendar_df <- read_csv(file_path)
```

In 'calendar_df' dataset, there are `r nrow(calendar_df)` rows and `r ncol(calendar_df)` columns in the dataset.
Columns names are  `r colnames(calendar_df)`.
```{r echo=FALSE, warning=FALSE}
#check the class of calendar_df
class(calendar_df)
```

```{r echo=FALSE, warning=FALSE}
#
nrow(calendar_df)

ncol(calendar_df)

colnames(calendar_df)
```

```{r echo=FALSE, warning=FALSE}
#check the column names for  calendar_df
colnames(calendar_df)
```

```{r echo=FALSE, warning=FALSE}
# examining the structure in calendar_df
str(calendar_df)
```

```{r echo=FALSE, warning=FALSE}
# examining the top 5 rows in calendar_df
head(calendar_df)
```

```{r echo=FALSE, warning=FALSE}
# examining summary of all numeric columns in calendar_df
summary(calendar_df)
```

#### Data Dictionary
| Field | Type  | Description |
|:-|:-|:-|
| listing_id | integer | AirBnB's unique identifier for the listing |
| date | date | Each Date of 365 days |
| available | text | TRUE means 'TRUE' and 'f' means 'FALSE' for that specific date |
| price | integer | Daily price in local currency |
| adjusted_price | integer | Daily price in local currency |
| minimum_nights | integer | Minimum nights for each AirBnB Listing |
| maximum_nights | integer | Maximum nights for each AirBnB Listing |

```{r echo=FALSE}
# examining the column type
str(calendar_df)
```

## Data Treatment
Column 'date' was changed to datetime type. A new column was created to extract the month of the dataset.

```{r echo=FALSE}
# converting to datetime object
calendar_df[['date']] <- as.POSIXct(calendar_df[['date']],format = "%Y-%m-%d")
```

```{r echo=FALSE}
str(calendar_df)
```

```{r echo=FALSE}
# extracting month from the date
calendar_df[['month']] <- format(calendar_df[['date']],format = "%B %Y")
```

```{r echo=FALSE, warning=FALSE}
# e3xamining unique months in the dataset
unique(calendar_df$month)
```

```{r  echo=FALSE}
head(calendar_df)
```


```{r  echo=FALSE}
min(calendar_df$date)
```

```{r  echo=FALSE}
max(calendar_df$date)
```

```{r echo=FALSE, warning=FALSE}
# changing 'month' to factor
calendar_df$month <- as.factor(calendar_df$month)

# changing  the order of factors
calendar_df$month <- factor(calendar_df$month, levels = c("June 2024", "July 2024", "August 2024", "September 2024","October 2024", "November 2024", "December 2024", "January 2025", "February 2025", "March 2025","April 2025", "May 2025", "June 2025"))
```

```{r  echo=FALSE}
#removing '$'
calendar_df$price <- substring(calendar_df$price, 2)
#removing '.00'
calendar_df$price <- str_sub(calendar_df$price, end = -4)

#removing '$'
calendar_df$adjusted_price <- substring(calendar_df$adjusted_price, 2)
#removing '.00'
calendar_df$adjusted_price <- str_sub(calendar_df$adjusted_price, end=- 4)

# changing column to integer
calendar_df$price <- strtoi(calendar_df$price)
# changing column to integer
calendar_df$adjusted_price <- strtoi(calendar_df$adjusted_price)
```

## Dropping JUne 2024
The dataset starts on the 26th June 2024, so data for June 2024 was dropped (Because only 4 days in June).
```{r echo=FALSE}
old_calendar_df_date <- min(calendar_df$date)
print(old_calendar_df_date)
```

```{r echo=FALSE}
calendar_df %>% 
  # Dropping June 2024 because it will skew the analysis 
  filter(available == TRUE & month != 'June 2024')
```

```{r echo=FALSE}
calendar_df <- calendar_df %>% 
  # dropping June 2024 because it will skew the analysis 
  filter( (available == TRUE) & (month != 'June 2024')) 
```


```{r echo=FALSE}
new_calendar_df_date <- min(calendar_df$date)
print(old_calendar_df_date)
print(new_calendar_df_date)

```


## Issues with Listings ID

Issues with LISTINGS Id were identified. THe first 7 digits are the important ones. 

```{r echo=FALSE}
calendar_df <- calendar_df %>% 
  mutate(listings = substr(listing_id, 1, 7))
```

## Dataset 'summarise_df'
A new dataset was created, where listings were examined by availability for each month within a year (26th June 2024 to 28th June 2025).
```{r echo=FALSE}
summarise_df <- calendar_df %>% 
  select(listing_id, month, available, price) %>% 
  # dropping June 2024 because it will skew the analysis 
  filter( (available == TRUE) & ('month' != 'June 2024')) %>% 
  group_by(listing_id, month) %>% 
  summarise(count=n(),
            min_price = min(price),
            mean_price = mean(price),
            median_price = median(price),
            max_price = max(price))
```

```{r echo=FALSE, warning=FALSE}
head(summarise_df)
```
```{r echo=FALSE, warning=FALSE}
# How many unique id?
print('summarise_df - unique ids')
print(n_distinct(summarise_df$listing_id))

print('summarise_df - ROWS')
print(nrow(summarise_df))
```

```{r echo=FALSE, warning=FALSE}
#
summarise_df %>%
  filter(count >=32)


```

#### Data Dictionary
| Field | Type  | Description |
|:-|:-|:-|
| listing_id | integer | AirBnB's unique identifier for the listing |
| month | date | Month of the year |
| count | integer | Number of nights available in that month |
| min_price | integer | Minimum price for that listing in that month |
| mean_price | integer | Mean price for that listing in that month |
| max_price | integer | Maximum price for that listing in that month |



In 'summarise_df' dataset, there are `r nrow(summarise_df)` rows and `r ncol(summarise_df)` columns in the dataset.
Columns names are `r colnames(summarise_df)`.

## Dataset 'available_nights_df
A new dataset summarising the number of available nights for each listing in a year (26th June 2024 to 28th June 2025).

```{r echo=FALSE, warning=FALSE}
# Examining availability for each host
available_nights_df <- calendar_df %>% 
  filter(available == TRUE) %>% 
  group_by(listing_id) %>% 
  summarise(available_nights=n(),
            min_price = min(price),
            mean_price = mean(price),
            median_price = median(price),
            max_price = max(price))

head(available_nights_df)

```
```{r echo=FALSE, warning=FALSE}
# How many unique id?
 
print(n_distinct(available_nights_df$listing_id))
print(nrow(available_nights_df))
```


#### Data Dictionary
| Field | Type  | Description |
|:-|:-|:-|
| listing_id | integer | AirBnB's unique identifier for the listing |
| available_nights | integer | Number of nights available during the year |
| min_price | integer | Minimum price per night |
| mean_price | integer | Mean price per night |
| median_price | integer | Median price per night |
| max_price | integer | Max price per night |


## Saving the summarise_df as csv file

```{r echo=FALSE, warning=FALSE}

###################
# Define the file path using here
file_path <- here(DIRECTORY, "summarise_df.csv")

# Save the dataframe as a CSV file
write.csv(summarise_df, file = file_path)

```

## Saving the available_nights_df as csv file

```{r echo=FALSE, warning=FALSE}

###################
# Define the file path using here
file_path <- here(DIRECTORY, "available_nights_df.csv")

# Save the dataframe as a CSV file
write.csv(summarise_df, file = file_path)

```

In 'available_nights_df' dataset, there are `r nrow(available_nights_df)` rows and `r ncol(available_nights_df)` columns in the dataset.
Columns names are `r colnames(available_nights_df)`.


```{r echo=FALSE, warning=FALSE}

```


# Data source
https://insideairbnb.com/get-the-data/
