==> code/Stage3_StoryTelling.Rmd <==
---
title: "Karina.Jonina.Project"
author: "KarinaJonina"
date: "5/1/2022"
output:
  html_document:
    df_print: paged
    toc: true
    theme: united
always_allow_html: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r echo=FALSE, warning=FALSE, message=FALSE}

library(tidyverse)
library(ggplot2)
library(dplyr)
library(stringr)
library(knitr)
library(plotly)
library(forcats)
library(crosstalk)
library(here)
library(camcorder)


DIRECTORY <- "data_2024"

# specifying the path to the CSV file
# file_path <- here(DIRECTORY, "airbnb_final_df_20214.csv")

file_path <- here(DIRECTORY, "airbnb_final_df_2024.csv")

# Read the CSV file
df <- read_csv(file_path)

# changing 'month' to factor
df$month_category <- as.factor(df$month_category)
# changing  the order of factors
df$month_category <- factor(df$month_category, levels = c("1 month", 
                                                            "2 months",
                                                            "3 months",
                                                            "4 months",
                                                            "5 months",
                                                            "6 months",
                                                            "7 months",
                                                            "8 months",
                                                            "9 months",
                                                            "10 months",
                                                            "11 months",
                                                            "12 months",
                                                            "More than 12 months"
                                                            ))



# dropping irrelevant columns
df <- df %>%
  select(-c(number_of_reviews,last_review, reviews_per_month,calculated_host_listings_count,
            number_of_reviews_ltm,license,minimum_nights))



# changing 'best_choice_airbnb_or_rent_per_year' to factor
df$best_choice_airbnb_or_rent_per_year <- as.factor(df$best_choice_airbnb_or_rent_per_year)
# changing  the order of factors
df$best_choice_airbnb_or_rent_per_year <- factor(df$best_choice_airbnb_or_rent_per_year, levels = c("AirBnB", 
                                                            "Rent"))


# specifying the path to the CSV file
file_path <- here(DIRECTORY, "airbnb_large_df_2024.csv")

# Read the CSV file
large_df <- read_csv(file_path)


# changing 'month' to factor
large_df$month <- as.factor(large_df$month)
# changing  the order of factors
large_df$month <- factor(large_df$month, levels = c("August 2024",
                                                          "September 2024",
                                                           "October 2024",
                                                          "November 2024",
                                                          "December 2024",
                                                          "January 2025",
                                                          "February 2025",
                                                          "March 2025",
                                                          "April 2025",
                                                          "May 2025",
                                                          "June 2025",
                                                          "July 2024",
                                                          "June 2024"   
                                                            ))

```

```{r echo=FALSE}
df %>%
  select(id, county,room_type, availability_365, price, airbnb_potential_annual_income, #rent_price_per_month, 
         rent_potential_annual_income, best_choice_airbnb_or_rent_per_year,annual_diff_rent_minus_Airbnb) 
```


```{r echo=FALSE}
######
#Occupancy, potential annual incomr from airbnb, differnece and best choice
###

df <- df %>%
  mutate( 
    
        # assuming that 70% of avaiable nights would be occupied
        occupancy_70 = round(availability_365 *0.7),
        
        # potential annual income based on 70% occupancy rate
        airbnb_potential_annual_income = occupancy_70 * price,
        
        # If POSITIVE, then AIRBNB brings more income
        # If NEGATIVE, then RENT brings more income
        annual_diff_rent_minus_Airbnb = airbnb_potential_annual_income - rent_potential_annual_income,
        
        best_choice_airbnb_or_rent_per_year = if_else(airbnb_potential_annual_income >= rent_potential_annual_income, 'AirBnB', 'Rent'))

```


```{r echo=FALSE}
# assigning colors to specific variable 
# will be used in graphs to ensure that the colours remain the same.
# creates association with colour and variables

#RPZ or Non-RPZ
pressure_colors = c( "#316395", "#D62728")

#Entire Unit, private room, shared room
room_colors = c('#54A24B', '#72B7B2', '#B279A2')

# rent or AirBnB
rent_colors = c('#565656', '#F58518')

#over-90 or under-90
threshold_colors = c('#FF9DA6', '#72B7B2')


# availability
reverse_viridis = c( "#FDE725FF",
                     '#B8DE29FF',
                     '#73D055FF',
                     '#29AF7FFF',
                     '#1F968BFF',
                     '#2D708EFF',
                     '#39568CFF',
                     '#482677FF',
                     "#440154FF")
```


## AirBnB OverView
``` {r}
df %>%
  group_by(room_type) %>%
    summarise(n = n_distinct(id)) %>%
  plot_ly(labels = ~room_type, values = ~n, type = 'pie',
          textinfo = 'label+percent',
          hoverinfo = 'label+value',
          marker = list(colors =room_colors),
          hole = 0.45) %>%
  layout(title = list(text = '<b>The Most Common Acccomodation Type in Ireland is <span style="color:#54A24B;">Entire Units</span></b>',
                      x = 0.5),
         showlegend = FALSE)

```

### Where are the AirBnB listings in Ireland?
```{r warning=FALSE, echo=FALSE}
# Calculate the total number of listings
total_listings <- df %>%
  summarise(total = n_distinct(id)) %>%
  pull(total)

# Create the plot with percentage labels
fig <- df %>%
  group_by(county) %>%
  summarise(n = n_distinct(id)) %>%
  mutate(
    county = fct_reorder(county, n, .desc = FALSE),
    percent = round((n / total_listings) * 100, 1)  # Calculate percentage
  ) %>%
  plot_ly(
    y = ~county, 
    x = ~n,
    text = ~paste0(percent, "%"),  # Add percentage labels
    textposition = 'outside',  # Position labels outside the bars
    hovertemplate = ~paste0("County: ", county, "<br>", "Number of AirBnB listings: ", n, "<br>", "Percentage: ", percent, "%")
  ) %>%
  add_bars(showlegend = FALSE) %>%
  layout(
    yaxis = list(title = FALSE, automargin = TRUE),  # Ensure y-axis labels fit
    xaxis = list(title = "# of AirBnB listings"),
    title = list(
      text = 'Which counties have the most AirBnB listings in Ireland?',
      x = 0.5,
      y = 0.95,
      xanchor = 'center',
      yanchor = 'top'
    ),
    margin = list(t = 50, b = 50, l = 50)# ,  # Adjust margins to fit the title and graph
   # height = 600  # Increase height to fit all counties
  )

fig

```

### When are most AirBnB listings Available?

```{r echo=FALSE, warning=FALSE}

sum_df <- large_df %>% 
  select(id, month) %>% 
  group_by(month) %>% 
  summarize(count_distinct = n_distinct(id))


sum_df %>%
  plot_ly(x = ~month, y = ~count_distinct, #color = ~rent_pressure_zone,
          hovertemplate = ~paste("Month:", month, "<br>", 
                                 "Number of AirBnB listings:", count_distinct)) %>%
  add_lines(showlegend = FALSE) %>%
  layout(xaxis = list(title = FALSE),
         yaxis = list(title = "# of AirBnB listings", range = c(0, max(sum_df$count_distinct))),
         title = "When are the most AirBnB listings Available?")

```

# Exploring Rent Pressure Zones
```{r echo=FALSE, warning=FALSE}
# creating a map of Ireland with AirBnB listings in RTZ differentiated 
g <- list(
  scope = 'europe',
  showland = TRUE,
  landcolor = toRGB("gray95"),
  subunitcolor = toRGB("gray85"),
  countrycolor = toRGB("gray85"),
  countrywidth = 0.5,
  subunitwidth = 0.5,

  resolution = 50,
  lonaxis = list(range = c(-12, -5)),
  lataxis = list(range = c(52, 55))
)



unique_id_df <- df %>%
  group_by(id, lea, rent_pressure_zone,council, room_type, latitude, longitude) %>%
  summarise(price = mean(price))

fig <- plot_geo(df, lat = ~latitude, lon = ~longitude, color = ~rent_pressure_zone)
fig <- fig %>% add_markers(marker = list(opacity = 0.2), showlegend=FALSE, colors = pressure_colors)
fig <- fig %>% layout(
    title = 'Which AirBnB listings are in the RTZs?', geo = g)

fig

```


### Summary of RTZs
```{r message=FALSE, echo=FALSE}
# creating a dataframe with rent_pressure_zone and room_type, arranging by count
df %>% 
  select(rent_pressure_zone, room_type, price) %>% 
  group_by(room_type, rent_pressure_zone) %>% 
  summarise(count = n(),
            
            mean_price = mean(price, na.rm = TRUE),
            max_price = max(price, na.rm = TRUE)) 

summary(df$price)
```

```{r echo=FALSE, warning=FALSE}

df %>%
  select(id, rent_pressure_zone) %>%
  group_by(rent_pressure_zone) %>%
  summarise(count_distinct = n_distinct(id))
```

```{r message=FALSE, echo=FALSE}
# creating a dataframe with rent_pressure_zone and room_type, arranging by count


df %>%
  group_by(rent_pressure_zone) %>%
  summarise(n = n_distinct(id)) %>%
  plot_ly(labels = ~rent_pressure_zone, values = ~n, type = 'pie',
          textinfo = 'label+percent',
          hoverinfo = 'label+value',
          hole = 0.5,
          marker = list(colors = pressure_colors)) %>%
  layout(title = list(text = '<b><span style="color:#D62728;">Nearly half</span> of all Airbnb listings are in <span style="color:#D62728;">Rent Pressure Zones</span>.</b>',
                      x = 0.5),
        legend = list(
          orientation = "h",
          y = -0.1,
          x = 0.5,
          xanchor = "center",
          yanchor = "bottom"
    ))

```

### Breakdown of AirBnB listings by Accomodation Type and RPZ
```{r echo=FALSE}



# Summarize the data and calculate percentages
df_summary <- df %>%
  group_by(rent_pressure_zone, room_type) %>%
  summarise(n = n_distinct(id)) %>%
  ungroup() %>%
  mutate(
    room_type = fct_reorder(room_type, n, .desc = TRUE),
    total = sum(n),
    percent = round((n / total) * 100, 1)
  )

    
    
df_summary %>%

  mutate(room_type = fct_reorder(room_type, n, .desc = TRUE)) %>%
  plot_ly(x = ~room_type, y = ~n, color = ~rent_pressure_zone,    
          text = ~paste(percent, "%"),  # Add percentage labels
    textposition = 'outside',  # Position labels outside the bars
          
          hovertemplate = ~paste(rent_pressure_zone, '-',  room_type, "<br>", 
                        "Number of AirBnB listings:", n)) %>%
  add_bars(showlegend=FALSE, colors = pressure_colors) %>%
  layout(xaxis = list(title = FALSE),
          yaxis = list(title = "# of AirBnB listings"),
         title = list(text = '<b>There are nearly twice as many Entire Units on Airbnb\n as Private Rooms in <span style="color:#D62728;">Rent Pressure Zones</span>.<b>',
                      x = 0.5),
    margin = list(t = 100, b = 50)
  )
   
```


```{r warning=FALSE, echo=FALSE}
df %>%
  plot_ly(x = ~room_type, y = ~price, color = ~rent_pressure_zone) %>%
  add_boxplot(marker = list(opacity = 0.2), showlegend = FALSE, colors = pressure_colors) %>% 
  layout(
    xaxis = list(title = FALSE),
    yaxis = list(title = "Price in Euro",
         zerolinecolor = '#ffff',
         zerolinewidth = 2,
         gridcolor = 'ffff',
         range = list(0, 450)), 
         title = list(text = '<b><span style="color:#D62728;">Entire Units in Rent Pressure Zones</span> are\n on average <span style="color:#D62728;">more expensive</span> than units outside of RPZs\n or other accommodation types</b>',
                      x = 0.5),
    boxmode = "group"  ) 
   
```

```{r echo=FALSE, warning=FALSE}

rpz_df <- df %>%
  filter(rent_pressure_zone == 'RPZ')
```


### Potential Annual Income
```{r echo=FALSE, warning=FALSE}
# geo styling
g <- list(
  scope = 'europe',
  showland = TRUE,
  landcolor = toRGB("gray95"),
  subunitcolor = toRGB("gray85"),
  countrycolor = toRGB("gray85"),
  countrywidth = 0.5,
  subunitwidth = 0.5,

  resolution = 50,
  lonaxis = list(range = c(-12, -5)),
  lataxis = list(range = c(52, 55))
)

fig <- plot_geo(rpz_df, lat = ~latitude, lon = ~longitude, color = ~best_choice_airbnb_or_rent_per_year)
fig <- fig %>% add_markers(marker = list(opacity = 0.4),colors = rent_colors,
                hovertemplate = ~paste("County:", county, "<br>",
                        "To Rent or To Buy:",best_choice_airbnb_or_rent_per_year,"<br>",
                        
                        "Potential Difference", annual_diff_rent_minus_Airbnb, "<br>",
                        
                        "Airbnb Potential Annual Income:", airbnb_potential_annual_income, "<br>",
                        "Rent Potential Annual Income:", rent_potential_annual_income
                        ))
fig <- fig %>% layout(
    title = list(
        text = '<b>Based on Potential Annual Income,\n<span style="color:#F58518;">Long-Term Leasing</span> would potentially bring as much as Airbnb</b>',
        x = 0.5,
        y = 0.95,
        yanchor = "top"
    ),
    legend = list(
        orientation = "h",
        y = -0.1,
        x = 0.5,
        xanchor = "center",
        yanchor = "bottom"
    ),
    margin = list(
        l = 50,  # left margin
        r = 50,  # right margin
        t = 100, # top margin
        b = 50   # bottom margin
    ),
    geo = g
)

fig
```


```{r echo=FALSE, warning=FALSE}

rpz_df %>%
  select(id, best_choice_airbnb_or_rent_per_year) %>%
  group_by(best_choice_airbnb_or_rent_per_year) %>%
    summarise(n = n_distinct(id))
```

```{r message=FALSE, echo=FALSE}
# Summarize the data
summary_best_choice_airbnb_or_rent_per_year <- rpz_df %>%
  group_by(best_choice_airbnb_or_rent_per_year) %>%
  summarise(n = n_distinct(id)) 

# Calculate the maximum value rounded to the nearest hundred
summary_best_choice_airbnb_or_rent_per_year_max <- round(max(summary_best_choice_airbnb_or_rent_per_year$n), -2)

# Create the pie chart
fig <- summary_best_choice_airbnb_or_rent_per_year %>%
  plot_ly(
    labels = ~best_choice_airbnb_or_rent_per_year, 
    values = ~n, 
    type = 'pie',
    textinfo = 'label+percent',
    hoverinfo = 'label+value',
    hole = 0.5,
    marker = list(colors = rent_colors)
  ) %>%
  layout(
    title = list(
      text = paste('<b><span style="color:#F58518;">Nearly', summary_best_choice_airbnb_or_rent_per_year_max, 
                   '</span> landlords would earn <span style="color:#F58518;">more income</span>\n by <span style="color:#F58518;">Long-Term Leasing</span> than Airbnb (70% occupancy) in RPZs.</b>'),
      x = 0.5
    ),
    legend = list(
      orientation = "h",
      y = -0.1,
      x = 0.5,
      xanchor = "center",
      yanchor = "bottom"
    ),
    margin = list(
      l = 50,  # left margin
      r = 50,  # right margin
      t = 100, # top margin
      b = 50   # bottom margin
    )
  )

# Print the plot
fig

```

 
```{r echo=FALSE}
# create new dataset where filtered by 'Entire Unit'
entire_rpz <-  rpz_df %>% 
  filter(room_type =='Entire rental unit')



entire_rpz_count <- entire_rpz %>%
  group_by(best_choice_airbnb_or_rent_per_year) %>%
    summarise(n = n_distinct(id)) 

#######################
#
#######################
# create new dataset where filtered by 'private 
private <-  rpz_df %>% 
  filter(room_type =='Private room')


private_rpz_count <- private %>%
  group_by(best_choice_airbnb_or_rent_per_year) %>%
  summarise(n = n_distinct(id)) 


# changing 'best_choice_airbnb_or_rent_per_year' to factor
private_rpz_count$best_choice_airbnb_or_rent_per_year <- as.factor(private_rpz_count$best_choice_airbnb_or_rent_per_year)
# changing  the order of factors
private_rpz_count$best_choice_airbnb_or_rent_per_year <- factor(private_rpz_count$best_choice_airbnb_or_rent_per_year, levels = c("AirBnB", 
                                                            "Rent"))

# changing 'best_choice_airbnb_or_rent_per_year' to factor
entire_rpz_count$best_choice_airbnb_or_rent_per_year <- as.factor(entire_rpz_count$best_choice_airbnb_or_rent_per_year)
# changing  the order of factors
entire_rpz_count$best_choice_airbnb_or_rent_per_year <- factor(entire_rpz_count$best_choice_airbnb_or_rent_per_year, levels = c("AirBnB", 
                                                            "Rent"))


#######################
# annual_diff_rent_minus_Airbnb
#######################

# If POSITIVE, then AIRBNB brings more income
# If NEGATIVE, then RENT brings more income

entire_rpz_more_income_rent <-  entire_rpz %>%
  filter(annual_diff_rent_minus_Airbnb <0) %>%
  mutate(annual_diff_rent_minus_Airbnb = abs(annual_diff_rent_minus_Airbnb))


private_room_rpz_more_income_rent <-  private %>%
  filter(annual_diff_rent_minus_Airbnb <0)  %>%
  mutate(annual_diff_rent_minus_Airbnb = abs(annual_diff_rent_minus_Airbnb))



```


```{r echo=FALSE}
fig <- plot_ly()
fig <- fig %>% add_pie(data = entire_rpz_count, labels = ~best_choice_airbnb_or_rent_per_year, values = ~n,
                                             #direction = 'counterclockwise',
                      direction = 'counterclockwise',
                      name = "Entire Unit", 
                      textinfo = 'label+percent+name',
                      hoverinfo = 'label+value+name',
                      marker = list(colors = rent_colors),
                      hole = 0.45,

                      domain = list(row = 0, column = 0))

fig <- fig %>% add_pie( data = private_rpz_count, labels = ~best_choice_airbnb_or_rent_per_year, values = ~n,
                                              #direction = 'counterclockwise',
                      direction = 'counterclockwise',
                      name = "Private Room", 
                      textinfo = 'label+percent+name',
                      hoverinfo = 'label+value+name',
                      marker = list(colors = rent_colors),
                      hole = 0.45, 
                      

                      
                      domain = list(row = 0, column = 1))

fig <- fig %>% layout(title = '<b><span style="color:#F58518;">More than half</span> of Entire Units and Private Rooms\nwould potentially bring more income by <span style="color:#F58518;">Long-Term Leasing</span> than Airbnb in RPZs.</b>',
              showlegend = F,
              grid=list(rows=1, columns=2),
              xaxis = list(showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE),
              yaxis = list(showgrid = FALSE, zeroline = FALSE, showticklabels = FALSE),
              
              annotations = list(
                list(
                text = '<b>Entire Unit</b>',
                x = 0.15,
                y = 0,
                xref = 'paper',
                yref = 'paper',
                showarrow = FALSE
                ),
                            list(
              text = '<b>Private Room</b>',
              x = 0.85,
              y = 0,
              xref = 'paper',
              yref = 'paper',
              showarrow = FALSE
              )
              )
              )

fig
```





```{r repeating_functions, echo=FALSE}


counties_with_most_income_from_leasing <- function(table, room_type_var) { 
  rpz_more_income_rent_county <- table %>%
    group_by(county) %>%
    summarise(n_id = n_distinct(id)) %>%
    mutate(county = fct_reorder(county, n_id, .desc = T)) 
  
  county_max_nid <- round(max(rpz_more_income_rent_county$n_id), -2)
  
  county_max <- rpz_more_income_rent_county %>%
    filter(n_id == max(n_id))%>%
    select(county) %>%
    pull()
  
  
  # Create the plot
  fig <- table %>%
    group_by(county, best_choice_airbnb_or_rent_per_year) %>%
    summarise(n_id = n_distinct(id)) %>%
    ungroup() %>%  # Ensure the factor reordering is applied correctly
    mutate(county = fct_reorder(county, n_id, .desc = FALSE)) %>%
    plot_ly(y = ~county, x = ~n_id, color = ~best_choice_airbnb_or_rent_per_year,
            hovertemplate = ~paste("County:", county, "<br>",
                                   "Accomodation Type:", best_choice_airbnb_or_rent_per_year, "<br>", 
                                   "Number of AirBnB listings:", n_id)) %>%
    add_bars(showlegend = FALSE, marker = list(color = '#F58518')) %>%
    layout(
      yaxis = list(title = FALSE, categoryorder = "total ascending"),  # Order y-axis by n_id
      xaxis = list(title = "# of AirBnB listings"),
      title = list(
        text = paste('<b>Based on Potential Annual Income,\nnearly',  county_max_nid , room_type_var, 'would bring more income\nby <span style="color:#F58518;">Long-Term Leasing</span> in' , county_max, '</b>'),
        x = 0.5,
        y = 0.95,
        xanchor = 'center',
        yanchor = 'top'
      ),
      margin = list(t = 100, b = 50)  # Adjust margins to fit the title and graph
    )
  
  return(fig)
}


function_rpz_more_income_rent_month_category_county_table <- function(table) { #
  rpz_more_income_rent_month_category_county_table <- table %>%
  group_by(month_category, county, id) %>%
  ungroup() %>%
  group_by(month_category, county)%>%
  summarise(
    n = n(),
    min_profit = min(annual_diff_rent_minus_Airbnb),
    mean_profit = mean(annual_diff_rent_minus_Airbnb), 
    median_profit = median(annual_diff_rent_minus_Airbnb),  
    max_profit = max(annual_diff_rent_minus_Airbnb)) 
  
  return(rpz_more_income_rent_month_category_county_table)
}


crosstable_function <- function(table) {
  entire_month_category_county_crosstable <- table %>%
    select(month_category, county, mean_profit) %>%
    pivot_wider(names_from = month_category, values_from = mean_profit)
  
  
  
  # Order by the '1 month' column in descending order
  entire_month_category_county_crosstable <- entire_month_category_county_crosstable %>%
    arrange(desc(`1 month`), desc = FALSE) %>%
    mutate(across(where(is.numeric), ~ round(., -1)))
  
  entire_month_category_county_crosstable[is.na(entire_month_category_county_crosstable)] <- 0
  
  # Create a heatmap
  heatmap_data <- as.matrix(select(entire_month_category_county_crosstable, -county))
  rownames(heatmap_data) <- entire_month_category_county_crosstable$county
  
  
  heatmap <- plot_ly(
    x = colnames(heatmap_data),
    y = rownames(heatmap_data),
    z = heatmap_data,
    type = "heatmap",
    colors = colorRamp(c("white", "#F58518"))
  ) %>%
    layout(
      title = "Average Potential Annual Earning Difference of Entire Units in RPZs by Month Category and County",
      xaxis = list(title = ""),
      yaxis = list(title = "")
    )
  
  
  # Add annotations
  for (i in seq_len(nrow(heatmap_data))) {
    for (j in seq_len(ncol(heatmap_data))) {
      heatmap <- heatmap %>%
        add_annotations(
          x = colnames(heatmap_data)[j],
          y = rownames(heatmap_data)[i],
          text = round(heatmap_data[i, j], 2),
          showarrow = FALSE,
          font = list(color = 'black')
        )
    }
  }
  
  # Show the plot
  return(heatmap)

}


```



```{r warning=FALSE, echo=FALSE}
counties_with_most_income_from_leasing(entire_rpz_more_income_rent, 'Entire Units' )


```


```{r}
entire_rpz_more_income_rent_month_category_county_table <- function_rpz_more_income_rent_month_category_county_table(entire_rpz_more_income_rent)#
crosstable_function(entire_rpz_more_income_rent_month_category_county_table)

```










```{r echo=FALSE}
counties_with_most_income_from_leasing(private_room_rpz_more_income_rent, 'Private Rooms' )

```

```{r warning=FALSE, echo=FALSE}

private_room_rpz_more_income_rent_month_category_county_table <- function_rpz_more_income_rent_month_category_county_table(private_room_rpz_more_income_rent)#
crosstable_function(private_room_rpz_more_income_rent_month_category_county_table)



```

# Appendix


### Final Dataset
In the final 'df' dataset, there are `r nrow(df) ` rows and `r ncol(df)` columns. Columns names are  `r colnames(df)`. 

Because our dataset was has been merged with 'available_nights_df', as described above, our dataset has AirBnB listings only available from 1st July 2024 to 25th July 2025. Any past AirBnB are not included in our dataset. 

```{r echo=FALSE}
summary(df)
```

```{r echo=FALSE}
ncol(df)
```

```{r echo=FALSE}
nrow(df)
```

```{r echo=FALSE}
n_distinct(df$id)
```


```{r echo=FALSE}
head(df)
```


```{r echo=FALSE, warning=FALSE}
# creating a map of Ireland with AirBnB listings
g <- list(
  scope = 'europe',
  showland = TRUE,
  landcolor = toRGB("gray95"),
  subunitcolor = toRGB("gray85"),
  countrycolor = toRGB("gray85"),
  countrywidth = 0.5,
  subunitwidth = 0.5,

  resolution = 50,
  lonaxis = list(range = c(-12, -5)),
  lataxis = list(range = c(52, 55))
)

fig <- plot_geo(df, lat = ~latitude, lon = ~longitude, color = ~county)
fig <- fig %>% add_markers(marker = list(opacity = 0.4), showlegend = FALSE,
                
                hovertemplate = ~paste("County:", county))
fig <- fig %>% layout(
    title = 'Where are the AirBnB listings in Ireland?', geo = g)

fig
```


```{r}
entire_rpz_more_income_rent %>%
  plot_ly(x = ~month_category, y = ~annual_diff_rent_minus_Airbnb) %>%
  add_boxplot(marker = list(opacity = 0.2), showlegend = FALSE, colors = '#565656') %>% 
  layout(
    xaxis = list(title = FALSE),
    yaxis = list(title = "Difference between Airbnb and Long-Term Leasing"#,
         # zerolinecolor = '#ffff',
         # zerolinewidth = 2,
         # gridcolor = 'ffff',
         # range = list(0, 450)
         ), 
         title = list(text = '<b>Landlords <span style="color:#565656;">Airbnb</span> short number of months make less money than those who Long-Term Lease</b> Earnings',
                      x = 0.5),
    boxmode = "group") 
   
```

```{r}

```
